requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
opencv-python-headless>=4.9.0
numpy>=1.26.0
pydantic>=2.7.0
pydantic-settings>=2.5.0
python-multipart>=0.0.9
aiohttp>=3.9.0
requests>=2.32.0
pyyaml>=6.0
